chore(intellisense): init refactor
