+++ src/intellisense/texDocumentFormatProvider.ts
@@ -60,4 +60,3 @@ export class TexDocFormatter implements vscode.DocumentFormattingEditProvider {
         ];
     }
 }
-
