+++ src/intellisense/texSymbolTreeProvider.ts
